requirements.txt
requests==2.31.0
python-dotenv==1.0.0
json5==0.9.14
pyyaml==6.0.1
rich==13.7.0
aiohttp==3.9.1
docker==6.1.3
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
urllib3==2.0.7
redis==5.0.1
